+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8.0 : TBD
 
 - **Maintenance**: Add tests for fields
+- **Change**: Remove DeclarativeFieldsMetaclass and import from Django instead.
 - **Change**: Msgpack dependency is no longer required.
 - **Change**: Empty values passed into a FormField now return {} rather than None.
 - **Fix**: Throw a more user friendly error when passing non-Enums to EnumField.